searchBox componented done.
